Revert "Linux msft wsl 4.19.y (#3)" (#6)
This reverts commit 271456763a85394e318b345744ad8fd2692678c2.

+++ include/linux/quota.h
@@ -263,7 +263,7 @@ enum {
 };
 
 struct dqstats {
-	unsigned long stat[_DQST_DQSTAT_LAST];
+	int stat[_DQST_DQSTAT_LAST];
 	struct percpu_counter counter[_DQST_DQSTAT_LAST];
 };
